added usePasteKeyframesItem to context menu helpers and added to Connector ui component
